last udpate

+++ file_proccessing/employee_class.h
@@ -6,7 +6,7 @@
 class Employee {
 public:
 	int id{}, age{};
-	long float salary{};
+	float salary{};
 	char name[10]{};
 };
